+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add new `Alert` component.
+- Feature: Add new `CloseButton` component.
+- Feature: Add new `CrossIcon` component.
+- Fix: Use `CloseButton` component instead of simple button in `Popover`.
+
 ## [8.1.0] - 2017-03-06
 
 Features: